pyproject.toml
[project]
name = "wildedge-demo-app"
version = "0.1.0"
description = "WildEdge demo: multi-model article pipeline with traditional ML, local LLM, and remote LLM"
requires-python = ">=3.10"
dependencies = [
    "wildedge-sdk",
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "onnxruntime",
    "transformers",
    "llama-cpp-python",
    "huggingface_hub",
    "numpy",
]

[tool.uv.sources]
wildedge-sdk = { path = "../wildedge-python", editable = true }

[dependency-groups]
dev = ["ruff"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
